DNM: Adjust file names to use parallelnode
The filenames need to be separate for the test processes that are
started so they don't overwrite each other. This needs to be done for
both the junit result files and the polarion result files.

Note: this commit contains changes in vendor folder, that need to be
replaced by a bump of kubevirt/qe-tools after PR https://github.com/kubevirt/qe-tools/pull/24
has been merged.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>