kubevirtci: move into it's own folder to allow OWNERS
We use [kubevirtci] for spinning up virtualized k8s clusters for
testing (locally and on CI). Therefore we source the cluster-up folder
into kubevirts' project folder root. Also we have the
cluster-up-sha.txt which is used to verify which version we use
currently, so that we can determine whether an update is necessary.

Related to how OWNERS files work, and since both of those items are
located in the root folder, we currently need a root approver for
approving pull requests that bump kubevirtci.

Thus, to be able for sig-buildsystem to own kubevirtci we move it to
it's own folder `kubevirtci`. In that folder we place an OWNERS file
with which we as sig-buildsystem claim ownership of kubevirtci.

[kubevirtci]: https://github.com/kubevirt/kubevirtci

Signed-off-by: Daniel Hiller <dhiller@redhat.com>